@@ -4,7 +4,7 @@ publish_to: 'none'
 version: 2.0.0
 
 environment:
-  sdk: '>=3.0.0'
+  sdk: ">=3.2.0 <4.0.0"
 dependencies:
   flutter:
     sdk: flutter
@@ -13,10 +13,10 @@ dependencies:
   flex_color_scheme: ^7.3.1
   colornames: ^0.1.0
   file_picker: ^5.5.0
-  path_provider: ^2.1.1
-  flutter_native_splash: ^2.3.2
-  states_rebuilder: ^6.3.0
-  uuid: ^3.0.7
+  path_provider: ^2.1.2
+  flutter_native_splash: ^2.3.10
+  states_rebuilder: ^6.4.0
+  uuid: ^4.3.3
   isar: ^4.0.0-dev.13
   isar_flutter_libs: ^4.0.0-dev.13
   hive: ^4.0.0-dev.2
@@ -25,7 +25,7 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^2.4.6
+  build_runner: ^2.4.8
   flutter_lints: ^3.0.0
   flutter_launcher_icons: ^0.13.1
   copy_with_extension_gen: ^5.0.4
