@@ -6,7 +6,7 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.1.5 <4.0.0'
+  sdk: ">=3.2.0 <4.0.0"
 dependencies:
   colornames: ^0.1.0
   flex_color_scheme: ^7.3.1
@@ -15,9 +15,9 @@ dependencies:
   freezed_annotation: ^2.4.1
   hive_flutter: ^1.1.0
   json_annotation: ^4.8.1
-  path_provider: ^2.1.1
-  states_rebuilder: ^6.3.0
-  uuid: ^4.2.2
+  path_provider: ^2.1.2
+  states_rebuilder: ^6.4.0
+  uuid: ^4.3.3
   path: ^1.8.3
   # file_picker: ^6.1.1
   appwrite: ^11.0.1
@@ -28,8 +28,8 @@ dependencies:
   
 
 dev_dependencies:
-  freezed: ^2.4.5
-  build_runner: ^2.4.6
+  freezed: ^2.4.7
+  build_runner: ^2.4.8
   json_serializable: ^6.7.1
   flutter_launcher_icons: "^0.13.1"
   flutter_test:
