@@ -4,27 +4,27 @@ publish_to: 'none'
 version: 0.1.0
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: ">=3.2.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   uuid: ^4.3.3
   file_picker: ^6.1.1
-  google_fonts: ^6.1.0
+  google_fonts: ^5.1.0
   colornames: ^0.1.0
-  path_provider: ^2.0.15
+  path_provider: ^2.1.2
   flex_color_scheme: ^7.3.1
   hive_flutter: ^1.1.0
   freezed_annotation: ^2.4.1
   json_annotation: ^4.8.1
   package_info_plus: ^5.0.1
-  states_rebuilder: ^4.4.0
+  states_rebuilder: ^6.4.0
   # manager:
     # git: https://github.com/kmcite2020/manager.git
 dev_dependencies:
   build_runner: ^2.4.8
-  freezed: ^2.4.6
+  freezed: ^2.4.7
   json_serializable: ^6.7.1
 
 
