@@ -3,26 +3,26 @@ description: A new Flutter project.
 publish_to: 'none'
 version: 1.0.0+1
 environment:
-  sdk: '>=3.1.2 <4.0.0'
+  sdk: ">=3.2.0 <4.0.0"
 dependencies:
   flutter:
     sdk: flutter
-  hive_flutter:
-  cupertino_icons: ^1.0.2
+  hive_flutter: ^1.1.0
+  cupertino_icons: ^1.0.6
   flex_color_scheme: ^7.3.1
   colornames: ^0.1.0
-  google_fonts: ^6.1.0
+  google_fonts: ^5.1.0
   file_picker: ^6.1.1
-  path_provider: ^2.1.1
-  uuid:
-  states_rebuilder: ^6.3.0
+  path_provider: ^2.1.2
+  uuid: ^4.3.3
+  states_rebuilder: ^6.4.0
   freezed_annotation: ^2.4.1
   json_annotation: ^4.8.1
   flutter_native_splash: ^2.3.10
 dev_dependencies:
   flutter_launcher_icons: ^0.13.1
-  build_runner: ^2.4.6
-  freezed: ^2.4.5
+  build_runner: ^2.4.8
+  freezed: ^2.4.7
   json_serializable: ^6.7.1
 flutter_native_splash:
   # This package generates native code to customize Flutter's default white native splash screen
