@@ -4,20 +4,20 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.1.2 <4.0.0'
+  sdk: ">=3.2.0 <4.0.0"
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.2
-  states_rebuilder: ^6.3.0
-  path_provider: ^2.1.1
+  cupertino_icons: ^1.0.6
+  states_rebuilder: ^6.4.0
+  path_provider: ^2.1.2
   flex_color_scheme: ^7.3.1
-  uuid: ^4.1.0
+  uuid: ^4.3.3
   colornames: ^0.1.0
-  hive_flutter: 
+  hive_flutter: ^1.1.0 
   intl: ^0.18.1
   freezed_annotation: ^2.4.1
-  json_annotation: 4.8.1
+  json_annotation: ^4.8.1
   flutter_bloc: ^8.1.3
   flutter_native_splash: ^2.3.10
 
@@ -25,8 +25,8 @@ dependencies:
 dev_dependencies:
   flutter_lints: ^2.0.0
   json_serializable: ^6.7.1
-  freezed: ^2.4.5
-  build_runner: ^2.4.6
+  freezed: ^2.4.7
+  build_runner: ^2.4.8
   flutter_launcher_icons: ^0.13.1
 
 flutter:
