@@ -4,21 +4,21 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.0.6 <4.0.0'
+  sdk: ">=3.2.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   file_picker: ^5.3.3
-  flex_color_scheme: ^7.2.0
-  path_provider: ^2.1.0
-  cupertino_icons: ^1.0.5
+  flex_color_scheme: ^7.3.1
+  path_provider: ^2.1.2
+  cupertino_icons: ^1.0.6
   english_words: ^4.0.0
   google_fonts: ^5.1.0
-  uuid: ^3.0.7
-  font_awesome_flutter: ^10.5.0
-  flutter_native_splash: ^2.3.1
-  states_rebuilder: ^6.3.0
+  uuid: ^4.3.3
+  font_awesome_flutter: ^10.7.0
+  flutter_native_splash: ^2.3.10
+  states_rebuilder: ^6.4.0
   colornames: ^0.1.0
   hive_flutter: ^1.1.0
   freezed_annotation: ^2.4.1
@@ -30,8 +30,8 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
   json_serializable: ^6.7.1
-  freezed: ^2.4.5
-  build_runner: ^2.4.6
+  freezed: ^2.4.7
+  build_runner: ^2.4.8
   flutter_lints: ^3.0.1
   lints: ^3.0.0
   built_value_generator:
