@@ -4,17 +4,17 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.1.2 <4.0.0'
+  sdk: ">=3.2.0 <4.0.0"
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.2
-  states_rebuilder: ^6.3.0
-  flutter_native_splash: ^2.3.3
-  google_fonts: ^6.1.0
-  uuid: ^4.1.0
+  cupertino_icons: ^1.0.6
+  states_rebuilder: ^6.4.0
+  flutter_native_splash: ^2.3.10
+  google_fonts: ^5.1.0
+  uuid: ^4.3.3
   flex_color_scheme: ^7.3.1
-  path_provider: ^2.1.1
+  path_provider: ^2.1.2
   json_annotation: ^4.8.1
   colornames: ^0.1.0
   hive_flutter: ^1.1.0
@@ -25,7 +25,7 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_launcher_icons: ^0.13.1
-  build_runner: ^2.4.6
+  build_runner: ^2.4.8
   json_serializable: ^6.7.1
 flutter:
   uses-material-design: true
