@@ -2,5 +2,4 @@ package dev.loner.chat_gpt
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity() {
-}
+class MainActivity: FlutterActivity()
